% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_pkg_name.R
\name{extract_pkg_name}
\alias{extract_pkg_name}
\title{Extract Package Name(s)}
\usage{
extract_pkg_name(
  x,
  pattern = "(?<=\\\\<u\\\\>\\\\*\\\\*`)(.*?)(?=`\\\\*\\\\*\\\\<\\\\/u\\\\>)",
  ...
)
}
\arguments{
\item{x}{\link[base]{character} \link[base]{vector}}

\item{pattern}{\link[base]{regex}}

\item{...}{additional parameters, currently not in use}
}
\description{
Extract package name(s) from a \link[base]{character} \link[base]{vector}
}
\examples{
extract_pkg_name('<u>**`R`**</u> package <u>**`MASS`**</u>')
extract_pkg_name(c('[R] package [MASS]', '[survival]'), pattern = '(?<=\\\\[)(.*?)(?=\\\\])')
# exception handling
extract_pkg_name(character())
extract_pkg_name('')
extract_pkg_name(NA_character_)
}
