% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd_.htest_array.R
\name{htest_array}
\alias{htest_array}
\alias{outer.cor.test}
\title{Array of \code{htest}}
\usage{
outer.cor.test(X, Y = X, ...)
}
\arguments{
\item{X, Y}{two \link[base]{matrix}es with same number of rows}

\item{...}{additional parameters of \link[stats]{cor.test}}
}
\value{
Function \link{outer.cor.test} returns a \link{htest_array} object.
}
\description{
..
}
\details{
..
}
\note{
This is a factorial structure of \code{xc}-by-\code{yc}
\emph{not} a pairwise \emph{combination} of \code{x}.
}
\examples{
# see ?rmd_.htest_array
}
