% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_.R
\name{render_}
\alias{render_}
\title{Create and Render R Markdown file}
\usage{
render_(
  content,
  path = tempdir(),
  document = c("html", "word", "pdf"),
  file = stop("must specify `file` explicitly"),
  autofold = (document == "html"),
  author = "tingting.zhan@jefferson.edu",
  rmd.rm = TRUE,
  ...
)
}
\arguments{
\item{content}{an R object}

\item{path}{\link[base]{character} scalar}

\item{document}{\link[base]{character} scalar, type of output document,
\code{'html'} (default), \code{'word'} or \code{'pdf'}}

\item{file}{\link[base]{character} scalar}

\item{autofold}{\link[base]{logical} scalar}

\item{author}{\link[base]{character} scalar, author's name and/or email}

\item{rmd.rm}{\link[base]{logical} scalar, whether to remove the generated \code{'.rmd'} file,
default \code{TRUE}}

\item{...}{..}
}
\description{
..
}
