% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0PACKAGE.R
\docType{package}
\name{rmarkdown.tzh-package}
\alias{rmarkdown.tzh}
\alias{rmarkdown.tzh-package}
\title{rmarkdown.tzh}
\description{
..
}
\author{
\strong{Maintainer}: Tingting Zhan \email{tingtingzhan@gmail.com}

}
\keyword{internal}
