% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd_.binTab.R
\name{rmd_.binTab}
\alias{rmd_.binTab}
\title{Create R Markdown Script for \link[DanielBiostatistics10th]{binTab} Object}
\usage{
\method{rmd_}{binTab}(x, xnm = substitute(x), ...)
}
\arguments{
\item{x}{a \link[DanielBiostatistics10th]{binTab}}

\item{xnm}{\link[base]{language} or \link[base]{character} scalar, call of \code{x}}

\item{...}{additional parameters, currently not in use}
}
\value{
Function \link{rmd_.binTab} returns a \link[base]{character} \link[base]{vector}.
}
\description{
Method dispatch to \link[DanielBiostatistics10th]{binTab} for S3 generic \code{rmd_} (in a different master package).
}
\keyword{internal}
