% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd_.DemographicTable.R
\name{rmd_.DemographicTable}
\alias{rmd_.DemographicTable}
\title{Create R Markdown Script for \link[DemographicTable]{DemographicTable}}
\usage{
\method{rmd_}{DemographicTable}(x, xnm, ...)
}
\arguments{
\item{x}{a \link[DemographicTable]{DemographicTable}}

\item{xnm}{\link[base]{character} scalar, call of \code{x}}

\item{...}{additional parameters, currently not in use}
}
\value{
Function \link{rmd_.DemographicTable} returns a \link[base]{character} \link[base]{vector}.
}
\description{
Method dispatch to \link[DemographicTable]{DemographicTable} for S3 generic \code{rmd_} (in a secret master package).
}
\examples{
library(DemographicTable)
list(
  '`DemographicTable`' = DemographicTable(CO2, groups = 'Type', include = c('conc', 'uptake'))
) |> render_(filename = 'DemographicTable')
}
