% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd_.htest_array.R
\name{rmd_.htest_array}
\alias{rmd_.htest_array}
\title{R Markdown Script of \link{htest_array}}
\usage{
\method{rmd_}{htest_array}(x, xnm, ...)
}
\arguments{
\item{x}{a \link{htest_array}}

\item{xnm}{\link[base]{character} scalar, call of \code{x}}

\item{...}{additional parameters, currently not in use}
}
\value{
Function \link{rmd_.htest_array} returns a \link[base]{character} \link[base]{vector}.
}
\description{
..
}
\examples{
set.seed(100)
x = matrix(rnorm(50), ncol = 5, dimnames = list(NULL, LETTERS[1:5]))
y = matrix(rnorm(30), ncol = 3, dimnames = list(NULL, letters[1:3]))
list(
  '`htest_array`' = outer.cor.test(x, y)
) |> render_(file = 'htest_array')
}
