% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd_.magick.R
\name{rmd_.magick-image}
\alias{rmd_.magick-image}
\title{R Markdown Script of \CRANpkg{magick} Image}
\usage{
\method{rmd_}{`magick-image`}(x, xnm, ...)
}
\arguments{
\item{x}{a \code{'magick-image'} object returned from function \link[magick]{image_read}}

\item{xnm}{..}

\item{...}{..}
}
\description{
R Markdown Script of \CRANpkg{magick} Image
}
\examples{
?magick::image_read
library(magick); list(
 '`magick-image`' = 'https://jeroen.github.io/images/frink.png' |> image_read()
) |> render_(file = 'magick')
}
