% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd_.maxT.R
\name{rmd_.maxT}
\alias{rmd_.maxT}
\title{rmd_.maxT}
\usage{
\method{rmd_}{maxT}(x, xnm, ...)
}
\arguments{
\item{x}{\link[mDFR]{maxT-class} object}

\item{xnm}{..}

\item{...}{..}
}
\description{
..
}
\examples{
\donttest{
library(mDFR)
?maxT_santos_test
ds = split(santos1, f = ~ Hr + antigen)
list(
 '`maxT`' = maxT_santos_test(data1 = ds$`18.CEF`, data0 = ds$`0.CEF`)
) |> render_(filename = 'maxT')
}
}
