% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd_.pairwise.htest.R
\name{rmd_.pairwise.htest}
\alias{rmd_.pairwise.htest}
\title{R Markdown Script of \code{pairwise.htest}}
\usage{
\method{rmd_}{pairwise.htest}(x, xnm, ...)
}
\arguments{
\item{x}{a \code{pairwise.htest} object, i.e., the returned value from functions
\link[stats]{pairwise.t.test}, \link[stats]{pairwise.wilcox.test}
or \link[stats]{pairwise.prop.test}}

\item{xnm}{\link[base]{character} scalar, call of \code{x}}

\item{...}{additional parameters, currently not in use}
}
\value{
Function \link{rmd_.pairwise.htest} returns a \link[base]{character} \link[base]{vector}.
}
\description{
..
}
\examples{
m = airquality |> 
  within.data.frame(expr = {
    Month = factor(Month, labels = month.abb[5:9])
  }) |>
  with(expr = pairwise.t.test(Ozone, Month, pool.sd = FALSE, p.adj = 'none'))
list(
  '`pairwise.htest`' = m
) |> render_(file = 'pairwise.htest')
}
