% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd_.patchwork.R
\name{rmd_.patchwork}
\alias{rmd_.patchwork}
\title{R Markdown Script of \link[patchwork]{patchwork}}
\usage{
\method{rmd_}{patchwork}(x, xnm, ncol = 2L, ...)
}
\arguments{
\item{x}{a \link[patchwork]{patchwork}}

\item{xnm}{..}

\item{ncol}{\link[base]{integer} scalar, see function \link[patchwork]{plot_layout}}

\item{...}{..}
}
\description{
R Markdown Script of \link[patchwork]{patchwork}
}
\note{
As for now (early 2025), tzh prefers package \pkg{patchwork} over
function \code{ggpubr::ggarrange}.
}
\examples{
packageDate('patchwork')
packageDate('ggpubr')
id = quote(list(
   All = TRUE, Female = (sex == 'Female'), Male = (sex == 'Male')
))
library(survival.tzh); list(
 '`patchwork`' = subset_ggKM(os ~ ph.ecog, subset = id, data = lung2)
) |> render_(file = 'patchwork')
}
