% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd_.R
\name{rmd_}
\alias{rmd_}
\alias{rmd_.htest}
\alias{rmd_.power.htest}
\alias{rmd_.data.frame}
\alias{rmd_.array}
\alias{rmd_.list}
\alias{rmd_.numeric}
\alias{rmd_.character}
\alias{rmd_.noquote}
\alias{rmd_.ggplot}
\alias{rmd_.gDesc}
\alias{rmd_.gList}
\alias{rmd_.flextable}
\title{R Markdown Script}
\usage{
rmd_(x, ...)

\method{rmd_}{htest}(x, xnm, ...)

\method{rmd_}{power.htest}(x, xnm, ...)

\method{rmd_}{data.frame}(x, xnm, ...)

\method{rmd_}{array}(x, xnm, ...)

\method{rmd_}{list}(x, xnm, ...)

\method{rmd_}{numeric}(x, ...)

\method{rmd_}{character}(x, ...)

\method{rmd_}{noquote}(x, xnm, ...)

\method{rmd_}{ggplot}(x, xnm, ...)

\method{rmd_}{gDesc}(x, xnm, ...)

\method{rmd_}{gList}(x, xnm, ...)

\method{rmd_}{flextable}(x, xnm, ...)
}
\arguments{
\item{x}{an R object}

\item{...}{additional parameters, currently not in use}

\item{xnm}{..}
}
\description{
R Markdown Script
}
\examples{
\donttest{
library(mDFR)
?maxT_santos_test
ds = split(santos1, f = ~ Hr + antigen)
list(
 '`maxT`' = maxT_santos_test(data1 = ds$`18.CEF`, data0 = ds$`0.CEF`)
) |> render_(file = 'maxT')
} # slow

set.seed(100)
x = matrix(rnorm(50), ncol = 5, dimnames = list(NULL, LETTERS[1:5]))
y = matrix(rnorm(30), ncol = 3, dimnames = list(NULL, letters[1:3]))
library(htest.tzh); list(
  '`htest_array`' = outer.cor.test(x, y)
) |> render_(file = 'htest_array')

library(rpart.tzh); library(rpart); list(
  '`rpart`' = rpart(Kyphosis ~ Age + Number + Start, data = kyphosis, model = TRUE)
) |> render_(file = 'rpart')

library(venn.tzh); list(
  '`venn`' = venn(list(A = state.name[1:20], B = state.name[15:30]))
) |> render_(file = 'venn')

library(consort.tzh); list(
  '`consort`' = consort_plot(data = dispos.data, 
    orders = c(trialno = 'Population', exclusion = 'Excluded', trialno = 'Allocated'),
    side_box = c('exclusion'), cex = 0.9)
) |> render_(file = 'consort')
list(
  '`htest`' = list(
   t.test(mpg ~ am, data = mtcars)
  ),
  '`power.htest`' = list(
   power.t.test(power = .90, delta = 1)
  )
) |> render_(file = 'htest')
list(
  '`data.frame`' = swiss
) |> render_(file = 'data.frame')
list(
  '`matrix`' = VADeaths
) |> render_(file = 'matrix')
library(ggplot2); list(
  '`ggplot`' = ggplot(mtcars, aes(wt, mpg)) + geom_point()
) |> render_(file = 'ggplot')
}
