% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd_.rpart.R
\name{rmd_.rpart}
\alias{rmd_.rpart}
\title{Create R Markdown Lines for \link[rpart]{rpart} Objects}
\usage{
\method{rmd_}{rpart}(x, xnm, ...)
}
\arguments{
\item{x}{\link[rpart]{rpart} object}

\item{xnm}{..}

\item{...}{..}
}
\value{
Function \link{rmd_.rpart} returns a \link[base]{character} \link[base]{vector}.
}
\description{
..
}
\examples{
library(rpart.tzh); library(rpart)
list(
  '`rpart`' = rpart(Kyphosis ~ Age + Number + Start, data = kyphosis, model = TRUE)
) |> render_(file = 'rpart')
}
