% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd_.matchit.R
\name{rmd_matchit}
\alias{rmd_matchit}
\alias{rmd_.summary.matchit}
\alias{rmd_.matchit}
\title{R Markdown Format of \link[MatchIt]{matchit} Object}
\usage{
\method{rmd_}{summary.matchit}(x, xnm, ...)

\method{rmd_}{matchit}(x, xnm, ...)
}
\arguments{
\item{x}{a \link[MatchIt]{matchit}}

\item{xnm}{..}

\item{...}{..}
}
\description{
..
}
\examples{
library(MatchIt)
m = matchit(treat ~ age+educ+race+nodegree+married+re74+re75, data = lalonde)
list(
 '`matchit`' = m
) |> render_(filename = 'matchit1')
list(
 '`matchit`' = m |> summary(addlvariables = 're78')
) |> render_(filename = 'matchit2')
}
