% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd_.bibentry.R
\name{url2doi}
\alias{url2doi}
\title{Correct \code{doi} in \code{url}}
\usage{
url2doi(x)
}
\arguments{
\item{x}{a \link[utils]{citation} and/or \link[utils]{bibentry}}
}
\value{
Function \code{\link[=url2doi]{url2doi()}} returns a \link[utils]{citation} and/or \link[utils]{bibentry}.
}
\description{
Correct \code{doi} in \code{url}
}
\examples{
'stringi' |> citation() # using doi field, correct
'texreg' |> citation() # doi in url field, not good!

# \pkg{texreg} has textVersion hard-coded in CITATION file
# therefore 'textVersion' cannot be updated
'texreg' |> citation() |> url2doi() # only a partial fix 

'robslopes' |> citation() |> toBibtex()
'robslopes' |> citation() |> url2doi() |> toBibtex() # complete fix!

}
\keyword{internal}
