% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_.R
\name{render_}
\alias{render_}
\title{Create and Render R Markdown files from an R project}
\usage{
render_(
  content,
  path = tempdir(),
  document = c("html", "word", "pdf"),
  filename = stop("must specify `filename` explicitly"),
  autofold = (document == "html"),
  author = "tingting.zhan@jefferson.edu",
  rmd.rm = TRUE,
  ...
)
}
\arguments{
\item{content}{an R object}

\item{path}{\link[base]{character} scalar}

\item{document}{\link[base]{character} scalar, type of output document,
\code{'html'} (default), \code{'word'} or \code{'pdf'}}

\item{filename}{\link[base]{character} scalar}

\item{autofold}{\link[base]{logical} scalar}

\item{author}{\link[base]{character} scalar, author's name and/or email}

\item{rmd.rm}{\link[base]{logical} scalar, whether to remove the generated \code{'.rmd'} file,
default \code{TRUE}}

\item{...}{..}
}
\description{
..
}
\examples{
library(flextable)
library(ggplot2)
out = list(
  A = 1:3,
  B = flextable(head(swiss)),
  C = ggplot(mtcars, aes(wt, mpg)) + geom_point(),
  D = list(
   t.test(mpg ~ am, data = mtcars),
   power.t.test(power = .90, delta = 1)
  )
)
render_(out, filename = 'test')

}
