% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd_.bibentry.R
\name{rmd_.bibentry}
\alias{rmd_.bibentry}
\title{R Markdown Script of \link[utils]{citation} and/or \link[utils]{bibentry}}
\usage{
\method{rmd_}{bibentry}(x, ...)
}
\arguments{
\item{x}{a \link[utils]{citation} and/or \link[utils]{bibentry}}

\item{...}{additional parameters, currently not in use}
}
\value{
Function \link{rmd_.bibentry} returns a \link[base]{character} scalar or \link[base]{vector}.
}
\description{
R markdown format of a \link[utils]{citation} and/or \link[utils]{bibentry} object.
}
\details{
Function \link{rmd_.bibentry} beautifies the output from
function \code{utils:::format.bibentry} (with option \code{style = 'text'})
in the following ways.
\itemize{
\item{Line break \code{'\\n'} is replaced by a white space;}
\item{Fancy quotes \eqn{``}, \eqn{''}, \eqn{`} and \eqn{'} are removed;}
\item{doi entries are shown as URLs with labels (in R markdown grammar).}
}
}
\examples{
citation() |> rmd_.bibentry()
library(ggplot2); 'ggplot2' |> citation() |> rmd_.bibentry()
list(
 'Methods' = list(
  'String manipulation by **`R`** package **`stringi`**.',
  'Conversion of R regression output to LaTeX or HTML by **`R`** package **`texreg`**.'
 ),
 '`bibentry`' = list(
  'stringi' |> citation(), # good doi
  'texreg' |> citation() # bad doi
 )
) |> render_(file = 'bibentry')

# compare
'stringi' |> citation() |> toBibtex() # using doi field, correct
'texreg' |> citation() |> toBibtex() # using url field, not good! have not dealt with this before

\dontrun{ # disabled for ?devtools::check
ct = installed.packages() |>
 rownames() |>
 lapply(FUN = function(i) i |> citation()) # slow

ct |> lapply(FUN = rmd_.bibentry)
}
}
