% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd_.bibentry.R
\name{rmd_.bibentry}
\alias{rmd_.bibentry}
\title{R Markdown Format of \link[utils]{citation} and/or \link[utils]{bibentry}}
\usage{
\method{rmd_}{bibentry}(x, ...)
}
\arguments{
\item{x}{a \link[utils]{citation} and/or \link[utils]{bibentry} object}

\item{...}{additional parameters, currently not in use}
}
\value{
Function \link{rmd_.bibentry} returns a \link[base]{character} scalar or \link[base]{vector}.
}
\description{
R markdown format of a \link[utils]{citation} and/or \link[utils]{bibentry} object.
}
\details{
Function \link{rmd_.bibentry} beautifies the output from
function \code{utils:::format.bibentry} (with option \code{style = 'text'})
in the following ways.
\itemize{
\item{Line break \code{'\\n'} is replaced by a white space;}
\item{Fancy quotes \eqn{``}, \eqn{''}, \eqn{`} and \eqn{'} are removed;}
\item{doi entries are shown as URLs with labels (in R markdown grammar).}
}
}
\examples{
citation() |> rmd_.bibentry()
library(ggplot2); 'ggplot2' |> citation() |> rmd_.bibentry()
if (FALSE) { # disabled for ?devtools::check
ap = rownames(installed.packages())
lapply(ap, FUN = function(i) i |> citation() |> rmd_.bibentry())
}
}
