% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd_.R
\name{rmd_}
\alias{rmd_}
\alias{rmd_.default}
\alias{rmd_.list}
\alias{rmd_.numeric}
\alias{rmd_.character}
\alias{rmd_.noquote}
\alias{rmd_.ggplot}
\alias{rmd_.gDesc}
\alias{rmd_.gList}
\alias{rmd_.flextable}
\title{Generate R Markdown Script}
\usage{
rmd_(x, ...)

\method{rmd_}{default}(x, xnm, ...)

\method{rmd_}{list}(x, xnm, ...)

\method{rmd_}{numeric}(x, ...)

\method{rmd_}{character}(x, ...)

\method{rmd_}{noquote}(x, xnm, ...)

\method{rmd_}{ggplot}(x, xnm, ...)

\method{rmd_}{gDesc}(x, xnm, ...)

\method{rmd_}{gList}(x, xnm, ...)

\method{rmd_}{flextable}(x, xnm, ...)
}
\arguments{
\item{x}{an R object}

\item{...}{additional parameters, currently not in use}

\item{xnm}{..}
}
\description{
Generate R Markdown Script
}
